requirements.txt
torch
transformers
opencv-python
accelerate
pillow
scipy
gradio
tenacity